Communication between frontend and backend
Changed protos messages: separated request and response due to the impossibility to separate request and response on the server.
Added a file containing all "final" data, so you just need to import the file to use variable inside it, instead of create them over all files.
Created an example.sh file that execute an example created to test the communication between the two clients using the server.